requirements.txt
baostock>=0.8.8
pandas>=1.3.0
numpy>=1.21.0
duckdb>=0.9.0
pyecharts>=2.0.0
jinja2>=3.0.0
pyyaml>=6.0
tqdm>=4.60.0
